Setup server_tls for pki infrastructure
The keys are stored on the host computer in a folder named pki_files.
The certificates should be generated on the host computer and then
placed into the folders, which are then bound to the containers via
docker compose.

This is also available to the client so that we can use the cert when
performing curl requests.