Changed the way client certificate and keys are mounted to client
Previously they were shared with the files on server_mtls. This was a
mistake. When we revoke the client certificate, it effectively deletes
the files due to the way it's linked to the client container. This
copies the files to a client folder, and then links it to the client
container.